+++ src/modules/settlements.ts
@@ -7,7 +7,7 @@ import { totals } from "./totals"
 
 export namespace settlements {
 
-    export function getOrCreateSettlement(txHash: Bytes, tradeTimestamp: BigInt, solver: Address, txGasPrice: BigInt): void { 
+    export function getOrCreateSettlement(txHash: Bytes, tradeTimestamp: i32, solver: Address, txGasPrice: BigInt): void { 
 
         let settlementId = txHash.toHexString()
         let network = dataSource.network()
